pyproject.toml
[project]
name = "GachaLogOnlineView"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "xtaodada", email = "xtao@xtaolink.cn"},
]
dependencies = [
    "httpx>=0.27.2",
    "fastapi>=0.114.0",
    "starlette>=0.38.4",
    "uvicorn>=0.30.6",
    "flet>=0.24.1",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "cashews>=7.3.0",
    "python-dotenv>=1.0.1",
    "ujson>=5.10.0",
    "apscheduler>=3.10.4",
]
requires-python = ">=3.10.0"
readme = "README.md"
license = {text = "AGPL"}


[tool.pdm]
distribution = false
